Develop

+++ back-end/Controller/FamilyGenealogy/FamilyMember.js
@@ -301,6 +301,7 @@ var addMember = async (req, res) => {
 
 var isBirthOrderExist = (memberID, birthOrder, listBirthOrderExist) => {
     console.log("Vào hàm isBirthOrderExist");
+    console.log(`listBirthOrderExist: ${listBirthOrderExist}`)
     console.log(`listBirthOrderExist length:  ${listBirthOrderExist.length}, memberID: ${memberID}, birthOrder: ${birthOrder}`);
     for (let i = 0; i < listBirthOrderExist.length; i++) {
         console.log(`listBirthOrderExist[i].BirthOrder: ${listBirthOrderExist[i].BirthOrder}, listBirthOrderExist[i].MemberID: ${listBirthOrderExist[i].MemberID}`);
@@ -369,6 +370,13 @@ var updateMember = async (req, res) => {
 
             return res.send(Response.dataNotFoundResponse());
         }
+        // nễu đã có người liên quan như vợ chồng hay con cái thì không thể thay đổi giới tính
+        if (await isHasRelatedPerson(dataMember[0]) == true) {
+            console.log("Đã vào trường hợp có người liên quan");
+            if (req.body.Male != dataMember[0].Male) {
+                return res.send(Response.badRequestResponse(null, "Thành viên này đã có người liên quan, không thể thay đổi giới tính"));
+            }
+        }
         // bắt đầu kiểm tra birthorder
         let listChild;
         // trường hợp có cha nhưng ko có mẹ 
@@ -405,6 +413,29 @@ var updateMember = async (req, res) => {
 
         return res.send(Response.internalServerErrorResponse());
     }
+}
+
+var isHasRelatedPerson = (dataMember) => {
+    return new Promise(async (resolve, reject) => {
+        try {
+            console.log("Vào hàm isHasRelatedPerson với dataMember: ", dataMember.MemberID);
+            let listChild = await FamilyManagementService.getMembersByParentID(dataMember.MemberID);
+            // console.log(`listChild: ${JSON.stringify(listChild)}`)
+            let listMarriage = await MarriageManagement.getMarriageByHusbandIDOrWifeID(dataMember.MemberID);
+            // console.log(`listMarriage: ${JSON.stringify(listMarriage)}`)
+            if (listChild.length > 0 || listMarriage.length > 0) {
+                console.log("Có người liên quan");
+                resolve(true);
+            } else {
+                console.log("Không có người liên quan");
+                resolve(false);
+            }
+
+        } catch (error) {
+            console.log(error)
+            reject(error)
+        }
+    })
 }
 
 var updateMemberToGenealogy = async (req, res) => {
@@ -501,11 +532,6 @@ var updateMemberToGenealogy = async (req, res) => {
 var deleteMember = async (req, res) => {
     try {
         db.connection.beginTransaction();
-        // Log ra thông tin trong req.query      
-        let dataMember = await FamilyManagementService.getMemberByMemberID(req.query.MemberID);
-        if (dataMember == null || dataMember.length == 0) {
-            return res.send(Response.dataNotFoundResponse());
-        }
         // các trường bắt buộc phải có trong req.query
         const requiredFields = [
             'MemberID'
@@ -518,7 +544,15 @@ var deleteMember = async (req, res) => {
             return res.send(Response.missingFieldsErrorResponse(missingFields));
         }
         console.log("No missing fields");
+        let dataMember = await FamilyManagementService.getMemberByMemberID(req.query.MemberID);
+        if (dataMember == null || dataMember.length == 0) {
+            return res.send(Response.dataNotFoundResponse());
+        }
+        console.log("dataMember: ", dataMember[0].MemberID);
+        await FamilyManagementService.deleteMemberRelated(dataMember[0]);
         await FamilyManagementService.deleteMember(req.query.MemberID);
+        // xóa ảnh cũ
+        await CoreFunction.deleteImage(dataMember[0].Image);
         dataRes = {
             MemberID: req.query.MemberID,
         }
